+++ example/lib/screens/settings/image_settings.dart
@@ -16,6 +16,5 @@ class ImageSettingsScreen extends StatelessWidget {
         title: ezL10n(context).isPageTitle,
         showSettings: false,
         body: const EzImageSettings(),
-        fab: EzBackFAB(context),
       );
 }