+++ example/lib/widgets/updater.dart
@@ -23,8 +23,8 @@ const String _github =
 const String _versionSource =
     'https://raw.githubusercontent.com/Empathetech-LLC/empathetech_flutter_ui/refs/heads/main/example/APP_VERSION';
 
-/// '2.2.1'
-const String _appVersion = '2.2.1';
+/// '2.2.2'
+const String _appVersion = '2.2.2';
 
 class EzUpdater extends StatefulWidget {
   /// Checks for Open UI updates
@@ -45,8 +45,10 @@ class _EzUpdaterState extends State<EzUpdater> {
 
   // Define custom functions //
 
-  /// Check for Open UI updates
+  /// Check for Open UI updates (Desktop only)
   void checkVersion() async {
+    if (isMobile()) return;
+
     final http.Response response = await http.get(Uri.parse(_versionSource));
 
     if (response.statusCode != 200) return;